requirements.txt
# 基础依赖
Flask==2.1.3
Werkzeug==2.1.2
requests==2.28.1
beautifulsoup4==4.11.1

# 数据分析
numpy==1.23.5
matplotlib==3.7.1
Pillow==9.4.0

# 深度学习框架
torch==2.0.1+cu118
torchvision==0.15.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# 生产部署
waitress==2.1.2
streamlit==1.23.1

# 可选依赖（Volc引擎API）
openai==1.3.0
